Move and rename config service to config lodaer service
