+++ src/app/core/config-loader/config-loader.service.ts
@@ -1,21 +1,19 @@
 import {Injectable} from '@angular/core';
 import {HttpClient} from '@angular/common/http';
-import {Config} from '../models';
-import {XmppService} from './xmpp';
-import {environment} from '../../environments';
+import {Config} from '../../models/';
+import {XmppService} from '../xmpp/';
+import {environment} from '../../../environments/';
 
 @Injectable()
-export class ConfigService {
+export class ConfigLoaderService {
   /**
    * The URL of the configuration file.
    * @type {string}
    */
   static readonly CONFIG_FILE = environment.config_url;
-  private readonly _config: Promise<Config>;
 
   constructor(private http: HttpClient,
               private xmppService: XmppService) {
-    this._config = this.loadConfig();
   }
 
   /**
@@ -23,12 +21,8 @@ export class ConfigService {
    * that is loaded on this services initialisation.
    * @returns {Promise<Config>}
    */
-  public getConfig(): Promise<Config> {
-    return this._config;
-  }
-
-  private loadConfig(): Promise<Config> {
-    return this.http.get(ConfigService.CONFIG_FILE)
+  public loadConfig(): Promise<Config> {
+    return this.http.get(ConfigLoaderService.CONFIG_FILE)
       .toPromise()
       .then(json => {
         const config = Config.fromJson(json);
