+++ src/app/core/config-loader/index.ts
@@ -1,0 +1,1 @@
+export * from './config-loader.service';
