+++ src/app/core/index.ts
@@ -4,6 +4,6 @@ export * from './header';
 export * from './notifications';
 export * from './xmpp';
 export * from './xmpp-feature-detection';
-export * from './config.service';
+export * from './config-loader/';
 export * from './global-error-handler.service';
 export * from './navigation.service';
